Refactoring + additional katas
