+++ generator/src/Generator/Kata/Training/Generator/CampaignStatistics/CampaignStatisticsGenerator.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Generator\Kata\Training\Generator\RomanNumbers;
+namespace Generator\Kata\Training\Generator\CampaignStatistics;
 
 use Generator\Kata\Training\Generator\AbstractTrainingGenerator;
 use Generator\Kata\Training\Generator\TrainingGeneratorInterface;
 
-class RomanNumbersGenerator extends AbstractTrainingGenerator implements TrainingGeneratorInterface
+class CampaignStatisticsGenerator extends AbstractTrainingGenerator implements TrainingGeneratorInterface
 {
-    const GENERATOR_NAME = 'Roman Numbers';
+    const GENERATOR_NAME = 'Campaign statistics';
 
     public function getName(): string
     {
