+++ generator/src/Generator/Kata/Training/Generator/FizzBuzz/KATA.md
