+++ generator/src/Generator/Kata/Training/Generator/SieveOfErathostenes/SieveOfErathostenesGenerator.php
@@ -2,30 +2,20 @@
 
 namespace Generator\Kata\Training\Generator\SieveOfErathostenes;
 
-use Generator\Kata\File\FileInterface;
-use Generator\Kata\Training\Dictionary\StandardTrainingFileGetterTrait;
+use Generator\Kata\Training\Generator\AbstractTrainingGenerator;
 use Generator\Kata\Training\Generator\TrainingGeneratorInterface;
-use Symfony\Component\Console\Output\OutputInterface;
 
-class SieveOfErathostenesGenerator implements TrainingGeneratorInterface
+class SieveOfErathostenesGenerator extends AbstractTrainingGenerator implements TrainingGeneratorInterface
 {
-    use StandardTrainingFileGetterTrait;
-
     const GENERATOR_NAME = 'Sieve Of Erathostenes';
-
-    public function generate(FileInterface $file, OutputInterface $output)
-    {
-        $file
-            ->setContent(
-                $this->getTrainingFileContent(__FILE__)
-            )
-            ->save();
-
-        $output->writeln(sprintf('<info>%s</info>', self::GENERATOR_MESSAGE));
-    }
 
     public function getName(): string
     {
         return self::GENERATOR_NAME;
+    }
+
+    protected function getTrainingFilePath(): string
+    {
+        return __FILE__;
     }
 }
